+++ tests/util.py
@@ -7,6 +7,7 @@ from decouple import config
 
 PATH = pathlib.Path(__file__).parent.absolute()
 ASSETS_PATH = os.path.join(PATH, 'assets')
+ARTIFACTS_PATH = os.path.join(PATH, 'artifacts')
 
 LOG_TESTING = config('LOG_TESTING', cast=bool, default=True)
 LOG = logging.getLogger('Testing')