+++ HISTORY.rst
@@ -1,8 +1,15 @@
 Changelog
 =========
 
-0.1.0 - ??.??.????
+0.1.0 - 02.03.2023
 ------------------
 
 Initial version
+
+0.2.0 - 19.03.2023
+------------------
+
+- Added experiment which trains a consensus ensemble of various MEGAN models which are trained on
+  different dataset splits and which also have different model parameters.
+